Adding interface for globus search
This module adds a relatively thin wrapper on the globus search API for ingesting function and metadata.  Same ingest function can also update metadata associated with a function.

Note: this requires the FuncX App Confidential Client to have access to the Globus Search Scope.  It does not use dependent tokens for simplicity although this could/should be added.