Adding EndpointQueue as higher abstraction
With the new Task objects, we don't want our RedisQueue's doing so much.  On top of this, RedisQueues provide too much transparency when they could basically function just as normal queues.  To that end, we have EndpointQueues which are just normal queues of Tasks, with the normal en/dequeue operations.