Incorporate initial pre-commit config
1. Add pre-commit config which runs flake8 with flake8-bugbear

There's one extremely minor flake8-bugbear lint failure to resolve
(renaming an unused loop iteration variable).

Additional hooks included:
- meta hooks to validate pre-commit's own config
- merge-conflict hook from the "main" pre-commit hook repo
- github workflow schema validator, which uses a schemastore
  jsonschema to protect against commits of a class of malformed
  workflow files

2. Add a contrib doc which covers test and lint setup

In particular, the contrib doc serves to explain/introduce pre-commit
sufficiently for everyone working on the repo to get it setup if they
so choose. Setting up the hooks with `pre-commit install` is
considered recommended but optional in the guide.

3. Tweak the github workflow to run pre-commit linting as early as
   possible

A lint-failing commit can typically be detected very quickly, relative
to other steps in a workflow. For example, there's no point installing
all of the dependencies of a project -- potentially a slow process --
if linting is going to fail anyway.

Because pre-commit handles the dependencies for its hooks
automatically as part of the framework, the `pre-commit` lint step can
run almost immediately, without depending on the python package
requirements.