Update to use funcx-common redis tools
The primary aim of this change is to use the funcx-common RedisPubSub
mechanism. However, this entails making changes to the Task class as
well (here renamed to RedisTask). The reason is that RedisTask must
conform to the TaskProtocol from funcx_common to be acceptable for use
with the pubsub from that library.

This could be done with a cast at the usage site (this is simplest),
but it's better to have things that will typecheck more cleanly in the
future. So take the opportunity to convert Task here to use
RedisField, HasRedisFieldsMeta, and so forth.