Refactor tests to use fakeredis
- Add fakeredis to test requirements
- Use fakeredis client instead of a bare mock for `mock_redis`
- Create a mock_redis_task_factory fixture which creates tasks
- Update tests to use this and pass

It is no longer necessary to mock over the task initializer at all
(the `from_id` method which is just a pointless copy of `__new__`).

Additional work which was necessary here:

Fix a bug that this surfaced: a missing call to `int()` on the
invocation counter. Includes a test to confirm the failure behavior
for non-int (bad) data.

Fix a bad test in testing task status -- 'result' and 'exception'
get removed from task data when they are `None`.

For now, the pubsub mock is left in place, as it is rarely used. This
should be changed to a pubsub backed by the same fakeserver at a later
date.