Improve test config using tox+make
`make test` runs the testsuite. `make lint` runs pre-commit linters.
`make py-safety` runs the safety check.
Under the hood, all of these are just tox invocations.

tox.ini config declares py37 as the main testenv python, declares
requirements, and marks all envs as `skip_install` (because the web
service is not yet declared as an installable package).

Coverage collection hooked via pytest-cov is included, and fixed
coverage config which declares some normal exclude_lines rules, along
with skip_missing (which is useful for focusing on uncovered code in
the reports).
Because the coverage invocation is now controlled (via `tox`), we
don't need to exclude things from reporting which are never being
collected.

GitHub Actions CI builds now use tox. This makes the build marginally
slower (packages need to be installed multiple times, once per tox
env), but it has the advantage that there are no steps which are
unique to CI. By requiring that the CI activity be declared as tox
envs, the activities are also runnable for local dev.
Once the requirements data is trimmed to be accurate, we can expect
this overhead to shrink down to a small, reasonable hit on our CI
times.