Update to the latest funcx-common and adapt usage
The change from taking hostname/port information to taking a redis URL
requires that the we plan for the future and format a redis URL with
the current config variables.

In the future, it should be possible to switch from from
REDIS_HOST/REDIS_PORT to having no configuration of the redis
connection handled directly by the webapp, and instead rely on the
default connection building being controllable via the
FUNCX_COMMON_REDIS_URL environment variable.
For this reason, a TODO comment is added which notes the removal of
these parameters from flask config.

Before this can be done, deployment logic in the helm charts needs to
be updated to set a redis URL via the env var there. If this is done
at the same time that the flask config variables are removed, the
application should "switch" between configuration modes.
Afterwards, support for these values in the application here can be
removed.