Begin testsuite cleanup
Don't use testclasses unnecessarily when running under pytest; just
use test functions. This is only applied to the "app init" unit test
for now.

When test data is small (a few lines), it's cleaner to keep it in the
test files themselves, written to a tmp file, than it is to manipulate
`__file__` to get a path to a nearby fixture file.

Setup the necessary fixtures in the top-level conftest to get an app
and test client automatically.

Remove an empty test folder, the 'models' test dir.

Begin dong away with a structure in which the test directory layout is
expected to match the concrete package layout. In practice, this
inevitably drifts over time and does not establish categories of
useful tests on its own, for use with `pytest tests/<dirname>`.