Use Auth State abstraction for authentication
Rewrite the main authentication logic in `authentication/auth.py` into
a clearer-cut abstration, the "AuthenticationState". An authentication
state is initialized from a token (typically parsed from an
Authorization header) and handles the callouts to Globus Auth and
records the resulting data in the state object.

This begins to organize the logic in the auth.py module:
- auth_state.py contains the AuthenticationState definition and the
  function which gets it from flask.g (potentially creating it)
- globus_auth.py contains functions for interacting with Globus Auth
  as a service, abstracting away details of the Auth API from the
  AuthenticationState
- auth.py still exists, and contains the remaining functionality (not,
  as-yet, organized) for handling authentication, including decorators
  and Globus Nexus/Groups callouts

The result is that we can rely on the application always fetching from
`g.auth_state` to see the currently authenticated user. The testsuite
then takes advantage of this by mocking over `flask.g` in a couple of
fixtures. One is a "raw" context manager (for full control), and one
is an automatic invocation of that context manager. This makes it
possible for a test to switch between authentications as different
users over the course of a single test.