Add more tests for AuthenticationState
This resolves necessary issues in config in order to allow the
AuthenticationState tests to function.

Specifically: the flask app is doing database initialization on first
request, which leads to errors when trying to test code outside of the
request-sending context. Refactor the handling of model loading and
init to allow the testsuite to call into sql DB init cleanly.

Add clarifying comments to some AuthenticationState tests.