Add tests for redis TaskGroup create, delete
Also fix an issue where `exists()` was returning an int, when a bool
is expected by surrounding code. `bool(exists(...))` solves this. Type
annotate `exists()` on RedisTask and TaskGroup to indicate the
expected return type.

Testing revealed odd (possibly unexpected) semantics for TaskGroup
when `user_id` is not set. There is now a test which demonstrates and
explains the behavior, but it also includes a TODO comment for
evaluating and considering changing it.