Add initial, minimal mypy config
Setup mypy as a tox target, add it to `make lint` and CI.
Configure mypy to be relatively non-strict. Many functions will not be
type-checked because they have no annotations, and several useful mypy
warning modes are not enabled.

This still flags several issues, in code that is not type-safe and
which has annotated signature lines. Resolve these issues with as few
changes as possible.