Introduce RedisS3Storage interface
This will only be active when `FUNCX_S3_BUCKET_NAME` is set in the
environment; otherwise the behavior of the application remains
unchanged.

Tests cover all usage within the 'task_storage' module, and ensure
that with the S3 env var set, we can read from S3. Any test can use
the mock_s3_bucket fixture to setup moto for mocked S3 testing.
If tests are reorganized to have a functional or integration testing
dir in the future, the moto fixture can be autoused within that
context.